+++ lib/active_interaction/base.rb
@@ -190,5 +190,11 @@ module ActiveInteraction
 
       fail Interrupt
     end
+
+    def self.inherited(subclass)
+      instance_var = :@_interaction_filters
+
+      subclass.instance_variable_set(instance_var, instance_variable_get(instance_var))
+    end
   end
 end